+++ test/model.test.ts
@@ -64,32 +64,6 @@ describe('Wollok model', () => {
       it('should return its siblings (omitting the same node)', () => {
         method.siblings().should.deep.equal([siblingMethod])
       })
-
-      it('should return existing next sibling for method', () => {
-        method.nextSibling()?.should.equal(siblingMethod)
-      })
-
-      it('should return undefined next sibling for last method', () => {
-        siblingMethod.nextSibling()?.should.be.not.ok
-      })
-
-    })
-
-    describe('previous siblings', () => {
-      const method = new Method({ name: 'm', parameters: [], isOverride: false, id: 'm1' })
-      const siblingMethod = new Method({ name: 'm2', parameters: [], isOverride: false, id: 'm2' })
-      const clazz = new Class({ name: 'C', supertypes: [], members: [method, siblingMethod], id: 'c1' })
-      method.parent = clazz
-      siblingMethod.parent = clazz
-
-      it('should return its previous siblings if it has', () => {
-        siblingMethod.previousSiblings().should.deep.equal([method])
-      })
-
-      it('should return an empty list if it does not have previous siblings', () => {
-        method.previousSiblings().should.deep.equal([])
-      })
-
     })
 
   })