+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Several non-breakable, no-argument methods where replaced with computed properties.
 - Dropped `isNode` function in favor of `instanceof Node` notation.
 - `cached` decorator now supports properties as well as methods.
+- Dropped exposed `Node` methods that asumed a sibling order.
 
 ## v3.1.1
 - Fixed bug where closures where unable to retain self reference when nested inside other closures.