+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - Dropped `defaultFieldValues` in favor of the cleaner `defaultValueFor` method.
 - Renamed `Variable.isGlobal` to the more accurate `isAtPackageLevel`.
 
+## v3.1.2
+- Moved local game code to language project.
+
 ## v3.1.1
 - Fixed bug where closures where unable to retain self reference when nested inside other closures.